Load ServerApp
