+++ src/py/flwr/server/__init__.py
@@ -16,7 +16,6 @@
 
 
 from . import driver, strategy
-from .app import ServerConfig as ServerConfig
 from .app import run_driver_api as run_driver_api
 from .app import run_fleet_api as run_fleet_api
 from .app import run_server_app as run_server_app
@@ -26,6 +25,7 @@ from .client_manager import ClientManager as ClientManager
 from .client_manager import SimpleClientManager as SimpleClientManager
 from .history import History as History
 from .server import Server as Server
+from .server_config import ServerConfig as ServerConfig
 from .serverapp import ServerApp as ServerApp
 
 __all__ = [
